+++ RELEASES.md
@@ -1,3 +1,9 @@
+- 1.9.0
+
+  - **MSRV**: Rust 1.56 or later is now required.
+
+  - The `hashbrown` dependency has been updated to version 0.12.
+
 - 1.8.2
 
   - Bump the `rustc-rayon` dependency, for compiler use only.